Removed ContentDrawers. Moved editor scripts.
The last idea was too smart. There was already a better way that
existed.

The editor scripts were moved, which caused some issues with referencing
Node classes with the NodeTypeAttribute. References have been moved over
to full name qualifiers.

This was done because Unity separates editor scripts from non-editor
scripts. The classes _should_ have been editor only, but were exposed so
they can be referenced. After OnDraw was added to NodeBase and NodeEdge,
the NodeBase class needed to move back into Editor space.

ContentDrawers were an attempt to circumvent this, but a lousy one. In
the future, NodeBase may be split into two classes again to prevent some
of this strangeness.